Moved the non module css files to stylesComponents without breaking too. I'm a bit scared to change them from css to react so how about we leave it at that?
