+++ src/app/stylesComponents/ThinkingAnimation.css
